Revert "removing compiler dependency from tsoptions package"
This reverts commit 8bcae0b2d18fd3d21f5e987e5749f083e61141e4, reversing
changes made to a14606f81ac1ae0e2ba170c32f2b50f8c4b1acb0.
