+++ internal/ast/utilities.go
@@ -1499,56 +1499,3 @@ func getExportAssignmentExpression(node *Node) *Node {
 func isAliasableExpression(e *Node) bool {
 	return IsEntityNameExpression(e) || IsClassExpression(e)
 }
-
-func IsComputedNonLiteralName(name *Node) bool {
-	return IsComputedPropertyName(name) && !IsStringOrNumericLiteralLike(name.Expression())
-}
-
-func IsQuestionToken(node *Node) bool {
-	return node != nil && node.Kind == KindQuestionToken
-}
-
-func TryGetTextOfPropertyName(name *Node) (string, bool) {
-	switch name.Kind {
-	case KindIdentifier, KindPrivateIdentifier, KindStringLiteral, KindNumericLiteral, KindBigIntLiteral,
-		KindNoSubstitutionTemplateLiteral:
-		return name.Text(), true
-	case KindComputedPropertyName:
-		if IsStringOrNumericLiteralLike(name.Expression()) {
-			return name.Expression().Text(), true
-		}
-	case KindJsxNamespacedName:
-		return name.AsJsxNamespacedName().Namespace.Text() + ":" + name.Name().Text(), true
-	}
-	return "", false
-}
-
-func forEachPropertyAssignment(objectLiteral *ObjectLiteralExpression, key string, callback func(property PropertyAssignment) *Node, key2 string) string {
-	if objectLiteral == nil {
-		return ""
-	}
-
-	for _, property := range (*objectLiteral).Properties.Nodes {
-		if !IsPropertyAssignment(property) {
-			continue
-		}
-		propName, _ := TryGetTextOfPropertyName(property.Name())
-		if key == propName || (key2 != "" && key2 == propName) {
-			callback(*property.AsPropertyAssignment())
-		}
-
-	}
-	return ""
-}
-
-func getTsConfigObjectLiteralExpression(tsConfigSourceFile *SourceFile) *ObjectLiteralExpression {
-	if tsConfigSourceFile != nil && tsConfigSourceFile.Statements.Nodes != nil {
-		expression := tsConfigSourceFile.Statements.Nodes[0].AsExpressionStatement().Expression
-		return expression.AsObjectLiteralExpression()
-	}
-	return nil
-}
-
-func ForEachTsConfigPropArray(tsConfigSourceFile *SourceFile, propKey string, callback func(property PropertyAssignment) *Node) string {
-	return forEachPropertyAssignment(getTsConfigObjectLiteralExpression(tsConfigSourceFile), propKey, callback, "")
-}
