+++ internal/compiler/checker.go
@@ -15356,7 +15356,7 @@ func (c *Checker) computeEnumMemberValues(node *ast.Node) {
 }
 
 func (c *Checker) computeEnumMemberValue(member *ast.Node, autoValue float64, previous *ast.Node) EvaluatorResult {
-	if ast.IsComputedNonLiteralName(member.Name()) {
+	if IsComputedNonLiteralName(member.Name()) {
 		c.error(member.Name(), diagnostics.Computed_property_names_are_not_allowed_in_enums)
 	} else {
 		text := member.Name().Text()
@@ -19545,7 +19545,7 @@ func (c *Checker) getContextualTypeForBindingElement(declaration *ast.Node, cont
 	if name == nil {
 		name = declaration.Name()
 	}
-	if ast.IsBindingPattern(name) || ast.IsComputedNonLiteralName(name) {
+	if ast.IsBindingPattern(name) || IsComputedNonLiteralName(name) {
 		return nil
 	}
 	parent := declaration.Parent.Parent
